restructure `Semantic/`
