+++ src/IMP2/Semantic/Structural.hs
@@ -1,7 +1,7 @@
 module IMP2.Semantic.Structural where
 
-import IMP2.Semantic.State
+import IMP2.State
 import IMP2.Syntax
 
 run :: (State, Stm) -> IMP State
-run = undefined
+run (st, Skip) = return st
